scripts/seed/pyproject.toml
[project]
name = "halalnomad-seed"
version = "0.1.0"
description = "Seed-data pipeline for HalalNomad — scrapes places from Google Places + cert directories, stages to Supabase for review, promotes to production places table."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.7.0",
    "rapidfuzz>=3.10.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
]
